Refactoring data type of price to BigDecimal

+++ retail-core/src/main/java/com/eu/retail/core/model/Product.java
@@ -3,15 +3,17 @@ package com.eu.retail.core.model;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Setter @Getter
 public abstract class Product {
     private int pnu;
     private String name;
-    private double price;
+    private BigDecimal price;
     private String desc;
 
 
-    public Product(int pnu, String name, double price, String desc){
+    public Product(int pnu, String name, BigDecimal price, String desc){
         this.pnu = pnu;
         this.name = name;
         this.price = price;
@@ -19,6 +21,6 @@ public abstract class Product {
     }
     public abstract boolean isWeighted();
 
-    public abstract double calculateTotal(double amount);
+    public abstract BigDecimal calculateTotal(double amount);
 
 }
